requirements.txt
psutil
tabulate
pandas
mysql-connector-python